Final task: - added service layer.

+++ application/src/main/java/com/senla/courses/repository/GroupWallMessageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupWallMessageRepository extends JpaRepository<GroupWallMessage, Long> {
+public interface GroupWallMessageRepository extends JpaRepository<GroupWallMessage, Integer> {
 }
